Split dali_backend.cc
