+++ src/CMakeLists.txt
@@ -1,6 +1,6 @@
 # The MIT License (MIT)
 #
-# Copyright (c) 2020 NVIDIA CORPORATION
+# Copyright (c) 2020-2021 NVIDIA CORPORATION & AFFILIATES
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy of
 # this software and associated documentation files (the "Software"), to deal in
@@ -26,6 +26,8 @@ add_subdirectory(dali_executor)
 add_library(
         triton-dali-backend SHARED
         dali_backend.cc
+        dali_model_instance.cc
+        dali_model.cc
         utils/triton.cc
 )
