+++ src/dali_backend.cc
@@ -1,6 +1,6 @@
 // The MIT License (MIT)
 //
-// Copyright (c) 2020 NVIDIA CORPORATION
+// Copyright (c) 2020-2021 NVIDIA CORPORATION & AFFILIATES
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -22,396 +22,12 @@
 
 #include <memory>
 
-#include "src/dali_executor/dali_executor.h"
-#include "src/dali_executor/io_buffer.h"
-#include "src/dali_executor/utils/dali.h"
-#include "src/dali_executor/utils/utils.h"
-#include "src/model_provider/model_provider.h"
-#include "src/utils/timing.h"
+#include "src/dali_model.h"
+#include "src/dali_model_instance.h"
 #include "src/utils/triton.h"
 #include "triton/backend/backend_common.h"
-#include "triton/backend/backend_model.h"
-#include "triton/backend/backend_model_instance.h"
 
 namespace triton { namespace backend { namespace dali {
-
-struct ModelParameters {
-  explicit ModelParameters(common::TritonJson::Value& model_config) {
-    model_config.MemberAsObject("parameters", &params_);
-  }
-
-  /**
-   * Rerurn a value of a parameter with a given `key`
-   * or `def` if the parameter is not present.
-   */
-  template<typename T>
-  T GetParam(const std::string& key, const T& def = T()) {
-    T result = def;
-    GetMember(key, result);
-    return result;
-  }
-
-  int GetNumThreads() {
-    return GetParam("num_threads", -1);
-  }
-
- private:
-  template<typename T>
-  void GetMember(const std::string& key, T& value) {
-    auto key_c = key.c_str();
-    if (params_.Find(key_c)) {
-      common::TritonJson::Value param;
-      TRITON_CALL_GUARD(params_.MemberAsObject(key_c, &param));
-      std::string string_value;
-      TRITON_CALL_GUARD(param.MemberAsString("string_value", &string_value));
-      value = from_string<T>(string_value);
-    }
-  }
-
-  common::TritonJson::Value params_;
-};
-
-class DaliModel : public ::triton::backend::BackendModel {
- public:
-  static TRITONSERVER_Error* Create(TRITONBACKEND_Model* triton_model, DaliModel** state);
-
-  virtual ~DaliModel() = default;
-
-  TRITONSERVER_Error* ValidateModelConfig() {
-    // We have the json DOM for the model configuration...
-    common::TritonJson::WriteBuffer buffer;
-    RETURN_IF_ERROR(model_config_.PrettyWrite(&buffer));
-    LOG_MESSAGE(TRITONSERVER_LOG_INFO,
-                (std::string("model configuration:\n") + buffer.Contents()).c_str());
-
-    return nullptr;  // success
-  }
-
-  const ModelProvider& GetModelProvider() const {
-    return *dali_model_provider_;
-  };
-
-  ModelParameters& GetModelParamters() {
-    return params_;
-  }
-
-
-  void ReadOutputsOrder() {
-    using Value = ::triton::common::TritonJson::Value;
-    Value outputs;
-    model_config_.MemberAsArray("output", &outputs);
-    for (size_t output_idx = 0; output_idx < outputs.ArraySize(); output_idx++) {
-      Value out;
-      std::string name;
-      outputs.IndexAsObject(output_idx, &out);
-      out.MemberAsString("name", &name);
-      output_order_[name] = output_idx;
-    }
-  }
-
-
-  const std::unordered_map<std::string, int>& GetOutputOrder() const {
-    return output_order_;
-  }
-
-
- private:
-  explicit DaliModel(TRITONBACKEND_Model* triton_model) :
-      BackendModel(triton_model), params_(model_config_) {
-    const char sep = '/';
-
-    const char* model_repo_path;
-    TRITONBACKEND_ArtifactType artifact_type;
-    TRITON_CALL_GUARD(
-        TRITONBACKEND_ModelRepository(triton_model_, &artifact_type, &model_repo_path));
-
-    std::stringstream dali_pipeline_path;
-    dali_pipeline_path << model_repo_path << sep << version_ << sep << GetModelFilename();
-    std::string filename = dali_pipeline_path.str();
-    LOG_MESSAGE(TRITONSERVER_LOG_INFO,
-                (make_string("Loading DALI pipeline from file ", filename).c_str()));
-    dali_model_provider_ = std::make_unique<FileModelProvider>(filename);
-  }
-
-  std::string GetModelFilename() {
-    std::string ret;
-    TRITON_CALL_GUARD(model_config_.MemberAsString("default_model_filename", &ret));
-    return ret.empty() ? "model.dali" : ret;
-  }
-
-  ModelParameters params_;
-  std::unique_ptr<ModelProvider> dali_model_provider_;
-  std::unordered_map<std::string, int> output_order_;
-};
-
-
-TRITONSERVER_Error* DaliModel::Create(TRITONBACKEND_Model* triton_model, DaliModel** state) {
-  TRITONSERVER_Error* error = nullptr;  // success
-  try {
-    *state = new DaliModel(triton_model);
-  } catch (const std::exception& e) {
-    LOG_MESSAGE(TRITONSERVER_LOG_ERROR, e.what());
-    error = TRITONSERVER_ErrorNew(TRITONSERVER_ERROR_UNKNOWN,
-                                  make_string("DALI Backend error: ", e.what()).c_str());
-  }
-
-  return error;
-}
-
-struct ProcessingMeta {
-  TimeInterval compute_interval{};
-  int total_batch_size = 0;
-};
-
-struct InputsInfo {
-  std::vector<IDescr> inputs;
-  std::vector<int> reqs_batch_sizes;  // batch size of each request
-};
-
-class DaliModelInstance : public ::triton::backend::BackendModelInstance {
- public:
-  static TRITONSERVER_Error* Create(DaliModel* model_state,
-                                    TRITONBACKEND_ModelInstance* triton_model_instance,
-                                    DaliModelInstance** state);
-
-  DaliExecutor& GetDaliExecutor() {
-    return *dali_executor_;
-  }
-
-  const DaliModel& GetDaliModel() const {
-    return *dali_model_;
-  }
-
-  void Execute(const std::vector<TritonRequest>& requests) {
-    DeviceGuard dg(GetDaliDeviceId());
-    TimeInterval exec_interval{};
-    start_timer_ns(exec_interval);
-    auto responses = CreateResponses(requests);
-    ProcessingMeta proc_meta{};
-    TritonError error{};
-    try {
-      proc_meta = ProcessRequests(requests, responses);
-    } catch (...) { error = ErrorHandler(); }
-    for (auto& response : responses) {
-      SendResponse(std::move(response), TritonError::Copy(error));
-    }
-    end_timer_ns(exec_interval);
-    for (auto& request : requests) {
-      ReportStats(request, exec_interval, proc_meta.compute_interval, !error);
-    }
-    ReportBatchStats(proc_meta.total_batch_size, exec_interval, proc_meta.compute_interval);
-  }
-
- private:
-  DaliModelInstance(DaliModel* model, TRITONBACKEND_ModelInstance* triton_model_instance) :
-      BackendModelInstance(model, triton_model_instance), dali_model_(model) {
-    auto serialized_pipeline = dali_model_->GetModelProvider().GetModel();
-    auto max_batch_size = dali_model_->MaxBatchSize();
-    auto num_threads = dali_model_->GetModelParamters().GetNumThreads();
-    DaliPipeline pipeline(serialized_pipeline, max_batch_size, num_threads, GetDaliDeviceId());
-    dali_executor_ = std::make_unique<DaliExecutor>(std::move(pipeline));
-  }
-
-  void ReportStats(TritonRequestView request, TimeInterval exec, TimeInterval compute,
-                   bool success) {
-    LOG_IF_ERROR(TRITONBACKEND_ModelInstanceReportStatistics(triton_model_instance_, request,
-                                                             success, exec.start, compute.start,
-                                                             compute.end, exec.end),
-                 "Failed reporting request statistics.");
-  }
-
-  void ReportBatchStats(uint32_t total_batch_size, TimeInterval exec, TimeInterval compute) {
-    LOG_IF_ERROR(TRITONBACKEND_ModelInstanceReportBatchStatistics(
-                     triton_model_instance_, total_batch_size, exec.start, compute.start,
-                     compute.end, exec.end),
-                 "Failed reporting batch statistics.");
-  }
-
-  /**
-   * @brief Create a response for each request.
-   *
-   * @return responses vector
-   */
-  std::vector<TritonResponse> CreateResponses(const std::vector<TritonRequest>& requests) {
-    std::vector<TritonResponse> responses;
-    responses.reserve(requests.size());
-    for (auto& request : requests) {
-      responses.push_back(TritonResponse::New(request));
-    }
-    return responses;
-  }
-
-  /**
-   * @brief Run inference for a given \p request and prepare a response.
-   * @return computation time interval and total batch size
-   */
-  ProcessingMeta ProcessRequests(const std::vector<TritonRequest>& requests,
-                                 const std::vector<TritonResponse>& responses) {
-    ProcessingMeta ret{};
-    auto inputs_info = GenerateInputs(requests);
-    start_timer_ns(ret.compute_interval);
-    auto outputs_info = dali_executor_->Run(inputs_info.inputs);
-    end_timer_ns(ret.compute_interval);
-    for (auto& bs : inputs_info.reqs_batch_sizes) {
-      ret.total_batch_size += bs;
-    }
-    auto dali_outputs =
-        AllocateOutputs(requests, responses, inputs_info.reqs_batch_sizes, outputs_info);
-    dali_executor_->PutOutputs(dali_outputs);
-    return ret;
-  }
-
-  /**
-   * @brief Generate descriptors of inputs provided by given \p requests
-   * @return input descriptors and batch size of each request
-   */
-  InputsInfo GenerateInputs(const std::vector<TritonRequest>& requests) {
-    uint32_t input_cnt = requests[0].InputCount();
-    std::vector<IDescr> inputs;
-    inputs.reserve(input_cnt);
-    std::unordered_map<std::string, IDescr> input_map;
-    std::vector<int> reqs_batch_sizes(requests.size());
-    for (size_t ri = 0; ri < requests.size(); ++ri) {
-      auto& request = requests[ri];
-      ENFORCE(request.InputCount() == input_cnt,
-              "Each request must provide the same number of inputs.");
-      for (uint32_t input_idx = 0; input_idx < input_cnt; ++input_idx) {
-        auto input = request.InputByIdx(input_idx);
-        auto input_byte_size = input.ByteSize();
-        auto input_buffer_count = input.BufferCount();
-        auto meta = input.Meta();
-        auto& idescr = input_map[meta.name];
-        for (uint32_t buffer_idx = 0; buffer_idx < input_buffer_count; ++buffer_idx) {
-          auto buffer = input.GetBuffer(buffer_idx, device_type_t::CPU, GetDaliDeviceId());
-          idescr.buffers.push_back(buffer);
-        }
-        if (idescr.meta.shape.num_samples() == 0) {
-          idescr.meta = meta;
-        } else {
-          ENFORCE(idescr.meta.type == meta.type,
-                  make_string("Mismatched type for input ", idescr.meta.name));
-          idescr.meta.shape.append(meta.shape);
-        }
-        if (input_idx == 0) {
-          reqs_batch_sizes[ri] = meta.shape.num_samples();
-        } else {
-          ENFORCE(meta.shape.num_samples() == reqs_batch_sizes[ri],
-                  "Each input in a request must have the same batch size.");
-        }
-      }
-    }
-    for (const auto& descrs : input_map) {
-      inputs.push_back(descrs.second);
-    }
-    return {inputs, reqs_batch_sizes};
-  }
-
-  int32_t GetDaliDeviceId() {
-    return !CudaStream() ? ::dali::CPU_ONLY_DEVICE_ID : device_id_;
-  }
-
-  /**
-   * @brief Allocate outputs expected by given \p requests.
-   *
-   * Lifetime of the created buffer is bound to each of the \p responses
-   * @param batch_sizes batch size of each request
-   */
-  std::vector<ODescr> AllocateOutputs(const std::vector<TritonRequest>& requests,
-                                      const std::vector<TritonResponse>& responses,
-                                      const std::vector<int>& batch_sizes,
-                                      const std::vector<OutputInfo>& outputs_info) {
-    assert(requests.size() > 0);
-    assert(requests.size() == responses.size());
-    assert(requests.size() == batch_sizes.size());
-    uint32_t output_cnt = requests[0].OutputCount();
-    for (auto& req : requests) {
-      ENFORCE(output_cnt == req.OutputCount(),
-              "All of the requests must expect the same number of outputs.");
-    }
-    ENFORCE(outputs_info.size() == output_cnt,
-            make_string("Number of outputs expected by the requests (", output_cnt,
-                        ") does not match the number of outputs from DALI pipeline (",
-                        outputs_info.size(), ")."));
-    const auto& output_indices = dali_model_->GetOutputOrder();
-    ENFORCE(output_cnt == output_indices.size(),
-            make_string("Number of outputs exptected by the requests (", output_cnt,
-                        ") does not match the number of outputs in the config (",
-                        output_indices.size(), ")."));
-    std::vector<ODescr> outputs(output_cnt);
-    outputs.reserve(output_cnt);
-    for (const auto& out_index : output_indices) {
-      auto name = out_index.first;
-      int output_idx = out_index.second;
-      auto shapes = split_list_shape(outputs_info[output_idx].shape, batch_sizes);
-      std::vector<OBufferDescr> buffers(requests.size());
-      IOMeta out_meta{};
-      out_meta.name = name;
-      out_meta.type = outputs_info[output_idx].type;
-      for (size_t ri = 0; ri < requests.size(); ++ri) {
-        out_meta.shape = shapes[ri];
-        auto output = responses[ri].GetOutput(out_meta);
-        buffers[ri] = output.AllocateBuffer(outputs_info[output_idx].device, GetDaliDeviceId());
-      }
-      out_meta.shape = outputs_info[output_idx].shape;
-      outputs[output_idx] = {out_meta, buffers};
-    }
-    return outputs;
-  }
-
-  TritonError ErrorHandler() {
-    TritonError error{};
-    try {
-      throw;
-    } catch (TritonError& e) {
-      LOG_MESSAGE(TRITONSERVER_LOG_ERROR, (e.what()));
-      error = std::move(e);
-    } catch (DaliBackendException& e) {
-      LOG_MESSAGE(TRITONSERVER_LOG_ERROR, (e.what()));
-      error = TritonError::Unknown(make_string("DALI Backend error: ", e.what()));
-    } catch (DALIException& e) {
-      LOG_MESSAGE(TRITONSERVER_LOG_ERROR, (e.what()));
-      error = TritonError::Unknown(make_string("DALI error: ", e.what()));
-    } catch (std::runtime_error& e) {
-      LOG_MESSAGE(TRITONSERVER_LOG_ERROR, (e.what()));
-      error = TritonError::Unknown(make_string("Runtime error: ", e.what()));
-    } catch (std::exception& e) {
-      LOG_MESSAGE(TRITONSERVER_LOG_ERROR, (e.what()));
-      error = TritonError::Unknown(make_string("Exception: ", e.what()));
-    } catch (...) {
-      LOG_MESSAGE(TRITONSERVER_LOG_ERROR, ("Unknown error"));
-      error = TritonError::Unknown("Unknown error");
-    }
-    return error;
-  }
-
-  std::unique_ptr<DaliExecutor> dali_executor_;
-  DaliModel* dali_model_;
-};
-
-
-TRITONSERVER_Error* DaliModelInstance::Create(DaliModel* model_state,
-                                              TRITONBACKEND_ModelInstance* triton_model_instance,
-                                              DaliModelInstance** state) {
-  TRITONSERVER_Error* error = nullptr;  // success
-  const char* instance_name;
-  RETURN_IF_ERROR(TRITONBACKEND_ModelInstanceName(triton_model_instance, &instance_name));
-
-  TRITONSERVER_InstanceGroupKind instance_kind;
-  RETURN_IF_ERROR(TRITONBACKEND_ModelInstanceKind(triton_model_instance, &instance_kind));
-
-  int32_t instance_id;
-  RETURN_IF_ERROR(TRITONBACKEND_ModelInstanceDeviceId(triton_model_instance, &instance_id));
-
-  try {
-    *state = new DaliModelInstance(model_state, triton_model_instance);
-  } catch (const std::exception& e) {
-    LOG_MESSAGE(TRITONSERVER_LOG_ERROR, e.what());
-    error = TRITONSERVER_ErrorNew(TRITONSERVER_ERROR_UNKNOWN,
-                                  make_string("DALI Backend error: ", e.what()).c_str());
-  }
-
-  return error;  // success
-}
 
 extern "C" {
